+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+
+- The `DefaultPredictor` class now automatically crops the input image
+  to dimensions that are a multiple of four.
+
 ## [v0.0.1]
 
 ### Changed
@@ -14,5 +21,6 @@ All notable changes to this project will be documented in this file.
 
 - Initial project files.
 
+[v0.0.2]: https://github.com/LEB-EPFL/DEFCoN-ImageJ/releases/tag/0.0.2
 [v0.0.1]: https://github.com/LEB-EPFL/DEFCoN-ImageJ/releases/tag/0.0.1
 [v0.0.0]: https://github.com/LEB-EPFL/DEFCoN-ImageJ/releases/tag/0.0.0